apps/chat/requirements.txt
# FastAPI и зависимости
fastapi==0.116.1
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.10

# OpenTelemetry - правильные версии для production
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-httpx==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-instrumentation-sqlalchemy==0.56b0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0

# Logging
structlog==24.4.0
python-json-logger==2.0.7

# LLM и интеграции - версии, совместимые с Pydantic v2
litellm==1.52.16
crewai==0.76.9
notion-client==2.2.1

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
types-PyYAML==6.0.12
types-requests==2.31.0.20240406